fixed unit test files missing

+++ Eindopdracht csharp/Eindopdracht unit tests/File Tests.cs
@@ -10,20 +10,29 @@ namespace Eindopdracht_unit_tests
         [TestMethod]
         public void TestSavingNewAccount()
         {
-            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
-            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
+            if (File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")))
+                File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
+
+            if (File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")))
+                File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
 
             DataSaver.AddNewClient("TestClient", "1234").GetAwaiter().GetResult();
             DataSaver.AddNewClient("TestClient2", "1234").GetAwaiter().GetResult();
             Assert.IsTrue(Directory.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient")), "Client directory was not made");
 
             Assert.IsTrue(DataSaver.ClientExists("TestClient"), "Client folder was not found");
+
+            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
+            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
         }
         [TestMethod]
         public void TestSavingMessage()
         {
-            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
-            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
+            if(File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")))
+                File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
+
+            if(File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")))
+                File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
 
             DataSaver.AddNewClient("TestClient", "1234").GetAwaiter().GetResult();
             DataSaver.AddNewClient("TestClient2", "1234").GetAwaiter().GetResult();
@@ -32,13 +41,11 @@ namespace Eindopdracht_unit_tests
 
             DataSaver.WriteMessageFile("TestClient2", "TestClient", time, "Here is a message").GetAwaiter().GetResult();
 
-            Console.WriteLine(File.ReadAllText(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")).ToString() + " - " + "TestClient2‎" + time + "‎Here is a message");
-            Console.WriteLine(File.ReadAllText(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")).ToString() + " - " + "TestClient2‎" + time + "‎Here is a message");
-
+            
             Assert.IsTrue(File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")), "Chat file not made");
             Assert.IsTrue(File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")), "Chat file not made");
-            Assert.IsTrue(String.Equals("TestClient2‎" + time + "‎Here is a message", File.ReadAllText(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")).ToString()), "message saved incorrectly");
-            Assert.IsTrue(String.Equals("TestClient2‎" + time + "‎Here is a message", File.ReadAllText(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")).ToString()), "message saved incorrectly");
+            Assert.AreEqual<string>("TestClient2‎" + time + "‎Here is a message", File.ReadAllText(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")).ToString(), "message saved incorrectly");
+            Assert.AreEqual<string>("TestClient2‎" + time + "‎Here is a message", File.ReadAllText(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")).ToString(), "message saved incorrectly");
 
             File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
             File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
@@ -46,8 +53,11 @@ namespace Eindopdracht_unit_tests
         [TestMethod]
         public void TestReadingMessage()
         {
-            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
-            File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
+            if (File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2")))
+                File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
+
+            if (File.Exists(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient")))
+                File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
 
             DataSaver.AddNewClient("TestClient", "1234").GetAwaiter().GetResult();
             DataSaver.AddNewClient("TestClient2", "1234").GetAwaiter().GetResult();
@@ -62,19 +72,12 @@ namespace Eindopdracht_unit_tests
 
             string[][] expectedAnswer = expectedAnswerList.ToArray();
 
-            Console.WriteLine(expectedAnswer[0][0] + " - " + DataSaver.GetMessageFile("TestClient", "TestClient2")[0][0]);
-            Console.WriteLine(expectedAnswer[0][1] + " - " + DataSaver.GetMessageFile("TestClient", "TestClient2")[0][1]);
-            Console.WriteLine(expectedAnswer[0][2] + " - " + DataSaver.GetMessageFile("TestClient", "TestClient2")[0][2]);
-            Console.WriteLine(expectedAnswer[1][0] + " - " + DataSaver.GetMessageFile("TestClient", "TestClient2")[1][0]);
-            Console.WriteLine(expectedAnswer[1][1] + " - " + DataSaver.GetMessageFile("TestClient", "TestClient2")[1][1]);
-            Console.WriteLine(expectedAnswer[1][2] + " - " + DataSaver.GetMessageFile("TestClient", "TestClient2")[1][2]);
-
-            Assert.IsTrue(expectedAnswer[0][0] == DataSaver.GetMessageFile("TestClient", "TestClient2")[0][0], "wrong chat");
-            Assert.IsTrue(expectedAnswer[0][1] == DataSaver.GetMessageFile("TestClient", "TestClient2")[0][1],"should be empty");
-            Assert.IsTrue(expectedAnswer[0][2] == DataSaver.GetMessageFile("TestClient", "TestClient2")[0][2],"should be empty");
-            Assert.IsTrue(expectedAnswer[1][0] == DataSaver.GetMessageFile("TestClient", "TestClient2")[1][0],"wrong sender");
-            Assert.IsTrue(expectedAnswer[1][1] == DataSaver.GetMessageFile("TestClient", "TestClient2")[1][1],"wrong time");
-            Assert.IsTrue(expectedAnswer[1][2] == DataSaver.GetMessageFile("TestClient", "TestClient2")[1][2],"wrong message");
+            Assert.AreEqual<string>(expectedAnswer[0][0], DataSaver.GetMessageFile("TestClient", "TestClient2")[0][0], "wrong chat");
+            Assert.AreEqual<string>(expectedAnswer[0][1], DataSaver.GetMessageFile("TestClient", "TestClient2")[0][1],"should be empty");
+            Assert.AreEqual<string>(expectedAnswer[0][2], DataSaver.GetMessageFile("TestClient", "TestClient2")[0][2],"should be empty");
+            Assert.AreEqual<string>(expectedAnswer[1][0], DataSaver.GetMessageFile("TestClient", "TestClient2")[1][0],"wrong sender");
+            Assert.AreEqual<string>(expectedAnswer[1][1], DataSaver.GetMessageFile("TestClient", "TestClient2")[1][1],"wrong time");
+            Assert.AreEqual<string>(expectedAnswer[1][2], DataSaver.GetMessageFile("TestClient", "TestClient2")[1][2],"wrong message");
 
             File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient", "TestClient2"));
             File.Delete(Path.Combine(DataSaver.GetDirectory(), "TestClient2", "TestClient"));
